Finish Montecarlopi
Finished up working on the Guile Montecarlopi program (at least, the first version of it), and got it working, and taking command line parameters.  So far, it looks to be reasonably efficient and effective.

Yet to investigate how (or even if "I should") to make true standalone executable programs, however.