Make Kn of Whispers work
Got a version going that basically works correctly.  Mostly.  I think.

The big remaining problem seems to be that it looks like the communicate calls aren't necessarily tail-calls for some reason...

Plus, I still have to fix up grid and ring, and there's a load of cruft throughout it now.  E.g. the use of the semaphore.