Complete first version of vectors
Completed a first stab at the processing in the vector processing function.

Currently, it does one vector addition, and one vector multiplication (which results in a full square matrix), and takes steps to bound the resultant numbers.

I'm not convinced this is the best implementation, but it's what I have come up with so far.