Finish first version of lingalg
That's not to say that it's actually good...  Just that it works right now.

I'm still unconvinced that the benchmarks as constructed are the best idea.

In particular, I doubt that using square matrices everywhere is a great idea, as probably they receive special optimisations, but I'm expecting mostly not to use square ones.