Racket (#1)
* Created Racket montecarlopi.  Needs to swapped from threads to places, though, to get parallelism.

* Changed real number operations to use flonums

I'm not sure if it has made the slightest difference in performance or not, to be honest...

Also updated the README a little.

* Added use of unsafe ops

Again, can't tell if it made any difference to the speed...

* Started on SelectTime

Still having some issues with it, however.

* Tweaked SelectTime

I modified SelectTime a little bit, to see if I could get it working.

I have solved the problems with void parameters, or whatever the complaint was, but am still having trouble with the sync over the channels.

* Changed SelectTime to 'option' style

Instead of both sides using an iteration counter, the receiver now simply waits for a 'NONE token.

This is similar to the approach used in places in Concurrent Programming in ML.

* Created Spawn

It works, but currently it only uses threads rather than places.  I will need to come back to this.

* Added Commstime

Based on some versions of Dr Chalmers one on his CPA Lang Shootout Github Repo.

It doesn't output the numbers currently, but it does appear to take longer when a larger iteration count is specified.

* Created Ring of Whispers

Currently it only does Ring, rather than Kn or Grid.

Furthermore, it doesn't seem to scale up all that amazingly well...

* Started adding Kn & Grid to whispers

Got a bit stuck though, since they need to do things to ensure they only send out one message, but receive precisely the correct number of messages...

* Continued working on kn of whispers

It still isn't working completely correctly - it appears that the put events aren't always removed from the list after they have been run, because for some reason they aren't always detected as equal when they come back as the result of the synchronisation.

Also updated the readme a little.

* Make Kn of Whispers work

Got a version going that basically works correctly.  Mostly.  I think.

The big remaining problem seems to be that it looks like the communicate calls aren't necessarily tail-calls for some reason...

Plus, I still have to fix up grid and ring, and there's a load of cruft throughout it now.  E.g. the use of the semaphore.

* Removed some cruft

From whispers.  Comments and now-extraneous prints.

* Implement Whispers-grid

I seem to have got the grid one working fine.  I think ring also will be fine, since I made chans functions for both list and vector.

Also, update the readme to reflect grid's differing input requirements.

* Complete first version of vectors

Completed a first stab at the processing in the vector processing function.

Currently, it does one vector addition, and one vector multiplication (which results in a full square matrix), and takes steps to bound the resultant numbers.

I'm not convinced this is the best implementation, but it's what I have come up with so far.

* Finish first version of lingalg

That's not to say that it's actually good...  Just that it works right now.

I'm still unconvinced that the benchmarks as constructed are the best idea.

In particular, I doubt that using square matrices everywhere is a great idea, as probably they receive special optimisations, but I'm expecting mostly not to use square ones.

Co-authored-by: James Cooper <30380166+jcoo092@users.noreply.github.com>