Create Makefile
Create a Makefile that handles the compilation process for Racket, and also automates the process of benchmarking each of the files.It is still a work-in-progress at this point though.  The benchmarking part isn't totally right as of yet.

It looks like there is an issue currently in that the method for getting the command line parameters into the program doesn't appear to work in compiled executable files, as compared running `.rkt` files with `racket`.

Still haven't actually introduced the use of places, either.