Change away from using the launcher option
It looks like the better bet when using places is to *avoid* the launcher option on `raco exe`.  When not using it, I don't see the error messages about main, and can use the regular compiled executable approach.  The performance, however, appears to be rather shockingly awful...  Will probably need to ask for some tips on how to improve that.