Finish Places upgrade for Monte Carlo Pi
This commit gets the Monte Carlo Pi program working with places, and ensures that it both distributes the threads reasonably evenly as well making sure that the exact number of threads to collect over matches exactly the number of threads issued to the places.

Also, update the Makefile a little.  The first update changes the number of iterations and threads to use, so that they now follow a consistent pattern of `n = (2**3)(n - 1)`.  The old individual build rules have also been replaced by two static rules that cover all the cases.  A bonus is that `make all` works again now.