Convert commstime.rkt to Places
Implement the conversion of the Communications Time benchmark over to use Places.  Three of the four main processes run on places, while the last one runs on the main thread.  I *think* this means the least overhead, but more importantly also means that the program doesn't terminate prematurely because the main thread has run out of things to do.