Change selecttime.rkt to use places
Instead of using two threads to run the program, two places are spawned, which then exchange messages between themselves as before.

Of note here is that a semaphore could no longer be used, as it wasn't something that could be sent on a place channel and thus wasn't accepted in `place/context`.  Instead, an extra place-channel pair is now spawned, and the main thread waits to sync on one end of that.