Make Whispers/Ring work with places
The Ring part of Whispers now works properly with Racket's places.  It creates a ring of places (with a function running on the main thread interposed between links at one place, to handle shrinking the itertaion count), and inside each place is a chain of channels.  A place receives a message on its rx place channel, puts it onto the channel at the start of its internal chain, picks the message back up again at the end of its internal chain, then passes it on to the next place via a transmission place channel.  When the iteration count reaches 0, each place passes the message on to its successor, then terminates.