Abandon conversion of Kn and Grid to places
The conversion of Kn and Grid to using places is proving to be quite complex, given that there needs to be some new wrangling of messages between threads via places.  This means that it is taking a long time - time which can't particularly be spared right now.  Thus, efforts to convert them to running with places have been abandoned, for now.  If Racket should prove to be competitive in the other tests, then these will be re-visited.