Racket place (#3)
* Create Makefile

Create a Makefile that handles the compilation process for Racket, and also automates the process of benchmarking each of the files.It is still a work-in-progress at this point though.  The benchmarking part isn't totally right as of yet.

It looks like there is an issue currently in that the method for getting the command line parameters into the program doesn't appear to work in compiled executable files, as compared running `.rkt` files with `racket`.

Still haven't actually introduced the use of places, either.

* Fix spaces to commas error

The functionality to transform lists of space-separated values to lists of comma separated values was not yet functional, meaning that some calls to hyperfine were incorrect.  This commit fixes that, by using the spaces-to-commas function correctly, with a `$(callspaces-to-commas,$(...))` invocation.

* Update REAMDE and Makefile

* Update REAMDE and Makefile again.  The Makefile is almost complete - it's just that there's a problem with naming some of the output benchmark files - specifically those which are run using nested for loops in the shell.

* Adapted montecarlopi.rkt for places

I have seemingly created a first working version of the Monte Carlo Pi simulation using Racket's places to gain some true parallelism.  I ended up using places on the basis of advice received from some very helpful people in the Racket Slack workspace.

Currently, this program has a few issues.  Chief among them is the fact that using the compiled program invocation style, i.e. `./executables/montecarlopi 48 12` or similar, doesn't work -- instead only the `racket -tm montecarlopi 48 12` style works currently.  Also, as of yet this program is yet to be adjusted to guarantee that an equal number of threads as expected by the receiver in the `collect-from-chan` function will be created in total across the places.

* Change away from using the launcher option

It looks like the better bet when using places is to *avoid* the launcher option on `raco exe`.  When not using it, I don't see the error messages about main, and can use the regular compiled executable approach.  The performance, however, appears to be rather shockingly awful...  Will probably need to ask for some tips on how to improve that.

* Finish Places upgrade for Monte Carlo Pi

This commit gets the Monte Carlo Pi program working with places, and ensures that it both distributes the threads reasonably evenly as well making sure that the exact number of threads to collect over matches exactly the number of threads issued to the places.

Also, update the Makefile a little.  The first update changes the number of iterations and threads to use, so that they now follow a consistent pattern of `n = (2**3)(n - 1)`.  The old individual build rules have also been replaced by two static rules that cover all the cases.  A bonus is that `make all` works again now.

* Convert commstime.rkt to Places

Implement the conversion of the Communications Time benchmark over to use Places.  Three of the four main processes run on places, while the last one runs on the main thread.  I *think* this means the least overhead, but more importantly also means that the program doesn't terminate prematurely because the main thread has run out of things to do.

* Change selecttime.rkt to use places

Instead of using two threads to run the program, two places are spawned, which then exchange messages between themselves as before.

Of note here is that a semaphore could no longer be used, as it wasn't something that could be sent on a place channel and thus wasn't accepted in `place/context`.  Instead, an extra place-channel pair is now spawned, and the main thread waits to sync on one end of that.

* Convert Spawn.rkt to places

Change Spawn to use places, and then create threads within each place.

* Tidy Spawn.rkt

* Modify Select Time to use  instead of an extra channel

* Make Whispers/Ring work with places

The Ring part of Whispers now works properly with Racket's places.  It creates a ring of places (with a function running on the main thread interposed between links at one place, to handle shrinking the itertaion count), and inside each place is a chain of channels.  A place receives a message on its rx place channel, puts it onto the channel at the start of its internal chain, picks the message back up again at the end of its internal chain, then passes it on to the next place via a transmission place channel.  When the iteration count reaches 0, each place passes the message on to its successor, then terminates.

* Tidy Whispers.rkt

* Abandon conversion of Kn and Grid to places

The conversion of Kn and Grid to using places is proving to be quite complex, given that there needs to be some new wrangling of messages between threads via places.  This means that it is taking a long time - time which can't particularly be spared right now.  Thus, efforts to convert them to running with places have been abandoned, for now.  If Racket should prove to be competitive in the other tests, then these will be re-visited.

* Tidy Whispers.rkt