Continue trying to debug selecttime.sml
I still haven't quite managed to figure out what the issue actually is, but I have determined that it is related to trying to use select over larger lists of events.  When simply using one particular event (as currently hardcoded into the system), everything works fine.  When trying to select over an arbitrary number of channels, however, there seems to be a non-deterministic (at least from my perspective as a CML user) problem where the CML system crashes rather than finishing properly.

I have found that this seems to become a clear issue when using 512+ iterations (I think I also tried with smaller, I don't remember), and 50 or more channels to select over.  Much under 50 and the program usually completes fully.  By 64 channels, it is almost guaranteed that the program won't complete.  You can tell whether it did or not, by seeing if the "Select Time completed" message is printed or not.  I also noticed, when printing out each message as it was received, that the greater the number of channels involved, the 'faster' the program would fail.  I.e.  fewer iterations were needed, on average, for the program to crash.  Testing with 512 channels suggested that the program would often crash within those first 12 iterations.

Unfortunately, I wasn't able to work out what the actual error that causes the crash is, despite trying to put exception handlers all over the program.  Looking at the CML source on GitHub, I can see that there is a randomisation function that appears to be used indirectly as part of the `CML.select` function.  I assume that some issue related to that causes these crashes, as it appears to be the only non-deterministic part of the system, but I couldn't manage to work out how everything pieced together.