M lton (#6)
* Created first go at MonteCarloPi.sml

* Second go.\n  This time, it compiles and I can launch CML, I just haven't worked out how to make the top-level program wait for all that to finish...\nAlso, I still need to add command line argument parsing.

* Remove executable from git.  Was accidentally committed before, but doesn't need to be there.

* Second go.

This time, it compiles and I can launch CML, I just haven't worked out how to make the top-level program wait for all that to finish...

Also, I still need to add command line argument parsing.

* Still haven't figured out how to make the program wait for the CML threads to finish.  I seem to have stumbled onto a way to make the compiler get stuck in what looks like it might be an infinite loop, however...

* Got program working correctly, thanks to input from Matt on Stack Overflow.  Just need to get the command line parameters working, now.

* Added support for command line parameters to montecarlopi.sml.  Also added a couple of targets to the Makefile for running benchmarks across the number of iterations, and number of threads, respectively.

* Changed program to using SML-NJ Util's Random module, but that doesn't seem to have done anything to help with multithreading.  It has, however, made the program muuuuuuch slower.  More than twice as slow, I estimate.

* Changed to using the RAND SML/NJ lib.  Seems to be even slower than RANDOM..., although actually probably slightly easier to use.

* Start on CommsTime.sml

* Partially complete CommsTime.sml.  While the structure is there, it seems to crash after the very first number has been consumed...

* Fix commstime.sml, based on Yawar's advice on the mailing list.  Turns out that I had mucked up passing in the channels to my successor function.

* Make further tweaks to commstime.sml per Yawar's comments on the MLton mailing list, in email dated 25 May 2020.

* Create first version of selecttime.sml.  This appears to have a bug though, in that I have observed that probably at least 9 times out of 10, the program stops printing out that is has received messages well before the end of the count (on 1024 iterations).  I think it must be crashing prior that.  Need to investigate.

* Continue trying to debug selecttime.sml

I still haven't quite managed to figure out what the issue actually is, but I have determined that it is related to trying to use select over larger lists of events.  When simply using one particular event (as currently hardcoded into the system), everything works fine.  When trying to select over an arbitrary number of channels, however, there seems to be a non-deterministic (at least from my perspective as a CML user) problem where the CML system crashes rather than finishing properly.

I have found that this seems to become a clear issue when using 512+ iterations (I think I also tried with smaller, I don't remember), and 50 or more channels to select over.  Much under 50 and the program usually completes fully.  By 64 channels, it is almost guaranteed that the program won't complete.  You can tell whether it did or not, by seeing if the "Select Time completed" message is printed or not.  I also noticed, when printing out each message as it was received, that the greater the number of channels involved, the 'faster' the program would fail.  I.e.  fewer iterations were needed, on average, for the program to crash.  Testing with 512 channels suggested that the program would often crash within those first 12 iterations.

Unfortunately, I wasn't able to work out what the actual error that causes the crash is, despite trying to put exception handlers all over the program.  Looking at the CML source on GitHub, I can see that there is a randomisation function that appears to be used indirectly as part of the `CML.select` function.  I assume that some issue related to that causes these crashes, as it appears to be the only non-deterministic part of the system, but I couldn't manage to work out how everything pieced together.

* Add spawn.sml.  It seems to be working fine.

* Add whispers.sml.  Grid has been implemented so far.

* Add working kn mode to whispers.sml

* Add working grid mode to whispers.sml

* Keep latest versions safe.

* Start on linalg.sml proper

* Complete vector part of linalg.sml.  I think.

* Complete linalg.sml, matching linalg.rkt.

* Correct Makefile to run benchmarks

* Update README and Makefile to reflect the reality of the repository as it is now.