Add Hyperfine scripts to repository and update documentation
Copies of the statistical analysis scripts that Hyperfine provides have been copied into this repository.  This is solely so that the precise scripts used are available in the future if anyone wants to reproduce the CML Benchmarks results.

This commit also modifies the Makefiles slightly, both to tweak wording in some comments and to modify the Racket Makefile so that you no longer need to prepend `./executables` to each program name when you want to `make` it.

Further, it updates the various READMEs to better reflect current practice, and corrects a few typos.