Correct Spawn.rkt
Change the behaviour of spawn.rkt to only create the places once per run of the program, instead of once per iteration.  This is much more akin to what will actually happen with the stereo matching scenarios, and arguably a much more fair comparison with MLton.

Also, this commit reduces the different iterations, so that they increase by 2**6 instead of 2**3, but still have the same upper limit.