Change TypedRacket's Makefile to a symlink
Since, to the best of my knowledge, the Makefiles for Racket and TypedRacket should actually be identical (the differences between the two flavours being confined to the source code files), I have swapped out the old TypedRacket Makefile for a soft symlink to Racket's Makefile.  This should ensure that the two always stay in sync with each other automatically, since, after all, one is just a reference to the real file.

Also, this commit corrects a problem with the current benchmarking for linalg and whispers, in that the file names for the output files weren't being adjusted based on the style in use, which had the effect of causing the last style used to overwrite the earlier styles.  This wasn't currently an issue for Whispers, since it only 'ring' enabled right now, but it meant that only results for 'mixed' were recorded for linalg.