Correct bug in SelectTime.rkt
Turns out that both Racket and Typed Racket had a fairly major bug in their Select Time implementations...  The part where they sync on the receiving channel was commented out.  This meant that essentially each side just fired all their messages into the channel, and then quit, without ever actually taking any out.  This had the effect of artificially making the benchmark appear to be about twice as fast as it should have been.

This commit fixes the error in the selecttime.rkt files, and brings the behaviour of the program back to what it is meant to be.